clk: rockchip: Set parent rate for DCLK_VOP clock on RK3228
Similar to RK3328, the DCLK_VOP on RK3228 is parented by the hdmiphy clk
and it is expected that the DCLK_VOP and hdmiphy clk rate are synced.

Use CLK_SET_RATE_PARENT and CLK_SET_RATE_NO_REPARENT flags, same as used
on RK3328, to make full use of all supported diplay modes.

Fixes: 0a9d4ac08ebc ("clk: rockchip: set the clock ids for RK3228 VOP")
Fixes: 307a2e9ac524 ("clk: rockchip: add clock controller for rk3228")
Signed-off-by: Jonas Karlman <jonas@kwiboo.se>