media: rkvdec: Add get_fmt_opaque and valid_fmt ops
Add a valid_fmt() ops to validate if a CAPTURE format is supported.

Expectation is that the valid_fmt() ops is a deterministic function and
the return value from get_fmt_opaque() ops is passed as input to the
valid_fmt() ops.

The CAPTURE format is reset when queue is not busy and format validation
fail.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>