net: phy: motorcomm: YT8011AN VDDIO voltage init from DTS
motorcomm 8011AN PHY is not capable of automatically detecting
the VDDIO voltage; therefore, it necessitates initialization
configurations that correspond to actual VDDIO voltage levels.

vddio property designates the PHY's IO voltage.

If the vddio property is not specified in the DTS,
the default 3V3 IO hardware initialization settings will be applied.

Change-Id: Iea056f30d8ad47e5970101b4543cfda6f5c843a9
Signed-off-by: Xu Xuehui <xxh@rock-chips.com>