drm/bridge: analogix_dp: Add &link_train.max_link_rate and &link_train.max_lane_count
The &link_train.link_rate and &link_train.lane_count are used as the
max_link_rate and max_lane_count to check the bandwidth limitation
, select output format and so on. However, they both also are used as
the actual link_rate and lane_count for the eDP controller and phy.

So we add &link_train.max_link_rate and &link_train.max_lane_count
in order to avoid the confusion in the use of above variables.

On the other hand, some unexpected errors will occur without this
patch in some cases.

When I test the PSR function of eDP panel module NE160QAM-NX1, the
link rate will be set to the 8.1G, which is the max supported link
rate of this module, after exiting the PSR state, but it should be
5.4G corresponding to 3840x2400p60 resolution.

analogix_dp_detect(): set the &link_train.link_rate to 8.1G
  -> analogix_dp_fast_link_train():
     use the &link_train.link_rate as the previous link config
    -> analogix_dp_set_link_bandwidth()
      -> phy_configure():
	 error: invalid link_rate config 8.1G

Change-Id: I1a4ae3c1eb308c9e0d25fb7736de52336ea7f880
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>