drm/rockchip: vop: Leave vblank enabled after entering the PSR mode
According to the following commits:
commit 9d0e3cac3517 ("drm/atomic: Allow vblank-enabled + self-refresh "disable"")
commit 2bdba9d4a3ba ("drm/rockchip: vop: Leave vblank enabled in self-refresh")

It is better to leave vblank enabled after entering the PSR mode, in
order to ensure that the vblank APIs (like DRM_IOCTL_WAIT_VBLANK) can
work in PSR mode. User space is not aware when the display system is
already in PSR mode, so this appears to be an API violation.

Without this patch, there may be a warning according to the following
check in drivers/gpu/drm/drm_atomic_helper.c:

static void
disable_outputs(struct drm_device *dev, struct drm_atomic_state *old_state)
{
......
		ret = drm_crtc_vblank_get(crtc);
		/*
		 * Self-refresh is not a true "disable"; ensure vblank remains
		 * enabled.
		 */
		if (new_crtc_state->self_refresh_active)
			WARN_ONCE(ret != 0,
				  "driver disabled vblank in self-refresh\n");
		else
			WARN_ONCE(ret != -EINVAL,
				  "driver forgot to call drm_crtc_vblank_off()\n");
		if (ret == 0)
			drm_crtc_vblank_put(crtc);
......
}

Change-Id: Ide735636ab53fd6977de5034b8cfc4c8341220af
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>