arm64: dts: rockchip: rk3576: Use scmi gpu clk as main GPU clock
CLK_GPU is the main clock for the GPU on RK3576, it's typical source
pll can be one of gpll, cpll, aupll, spll or lpll. For higher clock
rates it is also possible to use the gpu pvtpll as pll source.

The logic to switch between a normal pll and the pvtpll depending on
rate is handled in TF-A firmware, and exposed to Linux as a scmi clock.
TF-A will typically change to use normal pll for rates up to 200 MHz and
use pvtpll for 300 MHz or more.

Change to use the SCMI_CLK_GPU as the main GPU clock and the normal
CLK_GPU as a bus clk to model this in a similar way as on RK356x.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>