usb: dwc2: host: Ensure power on hcd if attempt power cycle
Test on RK3308 platform usb dwc2-host with a special SanDisk
Device (VID=0x0781, PID=0x5591), it can't be detected during
the system start phase with the following error log.

Moreseriously, when this problem occurs, the device cannot be
recognized even with hot plugging.

[    0.525833] usb 1-1: new high-speed USB device number 2 using dwc2
[    5.773864] usb 1-1: device descriptor read/64, error -110
[   16.525832] usb 1-1: device descriptor read/64, error -71
[   16.797824] usb 1-1: new high-speed USB device number 3 using dwc2
[   16.963826] usb 1-1: device descriptor read/64, error -71
[   17.237826] usb 1-1: device descriptor read/64, error -71
[   17.343901] usb usb1-port1: attempt power cycle

The root cause is that when the usb core do attempt power
cycle, the device is disconnected, so the dwc2 hcd driver
doesn't power on the controller. The patch makes sure power
on hcd during attempt power cycle.

Signed-off-by: William Wu <william.wu@rock-chips.com>
Change-Id: Ib4b5526b0320da2a45b9222f4f5ade629d4c7b00