net: dsa: realtek: fix reset-gpios handling
With CONFIG_RESET_GPIO enabled reset_control_get will fallback and look
for reset-gpios and creates a gpio reset control device around it.

This can interfere with gpiod_get and result in an EBUSY error.

Change to first look for reset-gpios and only fallback to look for a
reset control if none was found to fix the EBUSY error.

Fixes: 56998aa6b7f0 ("net: dsa: realtek: support reset controller")
Signed-off-by: Jonas Karlman <jonas@kwiboo.se>