drm/rockchip: dw-dp: compare the real bus format before mode change
The userspace can set color format or color depth as auto mode(driver
auto choose color format and color depth) or a specific mode(example,
set as RGB 8bit). In the case, Userspace change from the auto mode to
a specific mode, but the real color format and color depth used may be
no change. So it should not set the mode change flag as true to cause
a full modeset. The bus format is determined by the color format and
color depth. To compare the real bus format before set mode change flag
can avoid this unnecessary full modeset.

Change-Id: I73a655ae2f3fa0f49992d55161733a8047ad55cd
Signed-off-by: Zhang Yubing <yubing.zhang@rock-chips.com>