media: i2c: cameras: Fix sensor register sequence error in fastboot
The camera driver needs to be registered earlier than the ISP driver.
The registration of the ISP driver in the fast-start scenario is all
done through the subsys_initcall interface, which has a higher priority
than device_initcall_sync. If the camera driver uses device_initcall_sync,
in the dual-camera fast-start scenario, it may be registered later than
the ISP driver, resulting in the failure of the ISP link creation.

Signed-off-by: Yuefu Su <yuefu.su@rock-chips.com>
Change-Id: I87b783f78b72abbbff82ed334df9dad37799dbf2