arm64: dts: rockchip: rk3588-amp: update memory layout to fit MCU single uncacheable window
Since the RPMSG memory needs to be configured as uncacheable, and the
MCU only has a single uncacheable configuration window, it is necessary
to adjust the layout region to place the RPMSG adjacent to the register
region. This way, both the RPMSG and register addresses can be
configured as uncacheable simultaneously.

Signed-off-by: Zain Wang <wzz@rock-chips.com>
Change-Id: Ifc44a3ad5cd8a7206cb32bc6fa09b772de7ea13c