usb: gadget: u_audio: Fix req buf unaligned to cache line
The req buf is used for USB controller to transmit the
actual UAC data, and most of USB controllers support
DMA, so the req buf should be aligned to cache line to
make sure cache coherent.

The current code use kcalloc to allocate memory for
the req buf array, and use the prm->max_psize as the
element size of kcalloc which gets from the ep maxpsize,
and the ep maxpsize (aka. ep wMaxPacketSize ) maynot be
aligned to cache line because the value of ep maxpsize
is calculated based on parameters such as the sampling
rate and the number of channels of UAC in the handler
set_ep_max_packet_size().

For example, it the UAC1 is configured to:

- sample rate: 48KHz
- sample size: 24bit (3bytes)
- channel: 2
- HS ep bInterval: 4 (1ms)
- usb req_number: 4

Then the ep wMaxPacketSize = 2 * 3 * (48000/1000 + 1) = 294B

Buffer layout before this patch:

+==============+==============+==============+==============+
|   req buf0   |   req buf1   |   req buf2   |   req buf3   |
|   (294B)     |   (294B)     |   (294B)     |   (294B)     |
+==============+==============+==============+==============+

Offsets:
- req buf0: 0x0000 - 0x0125
- req buf1: 0x0126 - 0x024B
- req buf2: 0x024C - 0x0371
- req buf3: 0x0372 - 0x0497

Buffer layout after this patch (cacheline size = 64):

+==============+==============+==============+==============+
|   req buf0   |   req buf1   |   req buf2   |   req buf3   |
|   (320B)     |   (320B)     |   (320B)     |   (320B)     |
+==============+==============+==============+==============+

Offsets:
- req buf0: 0x0000 - 0x0139
- req buf1: 0x0140 - 0x027F
- req buf2: 0x0280 - 0x03BF
- req buf3: 0x03C0 - 0x04FF

Signed-off-by: William Wu <william.wu@rock-chips.com>
Change-Id: I9633aa926fbcf242c452d63048106632d95e5ab6